demo7/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
google-auth-oauthlib
msal
openpyxl
python-multipart
